+++ CuentasClaras/Api/Stats/NetworkBuyer.cs
@@ -9,11 +9,11 @@ namespace CuentasClaras.Api.Stats
     {
         public string BuyerId { get; set; }
         public string Name { get; set; }
-        public double TotalAmountUYU { get; set; }
+        public decimal TotalAmountUYU { get; set; }
         public string Type { get; set; }
 
         public string Id => $"BUYER-{BuyerId}";
-        public double Weight => TotalAmountUYU;
+        public decimal Weight => TotalAmountUYU;
 
     }
 }
