+++ CuentasClaras/Api/Stats/TopProductQuery.cs
@@ -11,8 +11,8 @@ namespace CuentasClaras.Api.Stats
         public string Description { get; set; }
         public string DescriptionLong { get; internal set; }
         public string UnitName { get; set; }
-        public double TotalAmountPerUnit { get; set; }
-        public double TotalAmountUYU { get; set; }
+        public decimal TotalAmountPerUnit { get; set; }
+        public decimal TotalAmountUYU { get; set; }
         public int Quantity { get; set; }
         public int UnitId { get; set; }
         public int ReleaseItemClassificationId { get; set; }
